Revert "Merge branch 'master' into feature/interpolation_node_heating_grid"
This reverts commit 573ba2f1d66cf896f805bb70b59a191801dccf94, reversing
changes made to bdbc0fe800850abb3127b3f65ec207b89a93792d.

+++ src/module/koeln/ReadFlash/FieldTypes_anari.h
@@ -98,5 +98,5 @@ struct ParticleField
   int nrpart;
   std::vector<partVec> particlePosition;
   std::vector<partVec> particleVelocity;
-  std::vector<double> particleMass;
+  std::vector<float> particleMass;
 };
